# apps/server/src/assets/api-openapi.yaml
openapi: 3.1.0
info:
  title: Trilium Notes Internal API
  version: 0.98.0
  description: |
    This is the internal API used by the Trilium Notes client application.
    
    **Important:** This API is primarily intended for internal use by the Trilium client.
    For external integrations, please use the [ETAPI (External Trilium API)](https://github.com/TriliumNext/Trilium/blob/main/docs/User%20Guide/User%20Guide/Advanced%20Usage/ETAPI%20(REST%20API).md#etapi-rest-api) instead.
    
    ## Authentication
    
    Most endpoints require session-based authentication. You can authenticate using:
    - **Password login**: POST to `/api/login` with password
    - **Token authentication**: Generate a token via `/api/login/token`
    - **Sync authentication**: Use document secret for sync operations
    
    ## Rate Limiting
    
    Authentication endpoints are rate-limited to prevent brute force attacks.
    
    ## CSRF Protection
    
    State-changing operations require CSRF tokens when using session authentication.
    
  contact:
    name: TriliumNext Issue Tracker
    url: https://github.com/TriliumNext/Trilium/issues
  license:
    name: GNU Affero General Public License v3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Default local server
  - url: https://your-trilium-server.com
    description: Your Trilium server
...
security:
  - sessionAuth: []
  - tokenAuth: []
paths:
...
  /api/notes/{noteId}:
    get:
      tags: [Notes]
      summary: Get note metadata
      operationId: getNote
      parameters:
        - $ref: '#/components/parameters/noteId'
      responses:
        '200':
          description: Note metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '404':
          description: Note not found
          
    delete:
      tags: [Notes]
      summary: Delete note
      operationId: deleteNote
      parameters:
        - $ref: '#/components/parameters/noteId'
        - name: taskId
          in: query
          required: true
          schema:
            type: string
        - name: eraseNotes
          in: query
          schema:
            type: boolean
            default: false
        - name: last
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '204':
          description: Note deleted
        '404':
          description: Note not found
...
  /api/notes/{noteId}/revision:
    post:
      tags: [Notes, Revisions]
      summary: Force save revision
      operationId: saveRevision
      parameters:
        - $ref: '#/components/parameters/noteId'
      responses:
        '201':
          description: Revision saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: trilium.sid
      description: Session cookie obtained from login
      
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: API token in format "Bearer {token}"
  parameters:
    noteId:
      name: noteId
      in: path
      required: true
      schema:
        type: string
      description: 12-character note ID
  schemas:
    Note:
      type: object
      required:
        - noteId
        - title
        - type
        - mime
        - isProtected
      properties:
        noteId:
          type: string
          description: 12-character note ID
          example: ur11rSfHkzeV
        title:
          type: string
          description: Note title
        type:
          $ref: '#/components/schemas/NoteType'
        mime:
          type: string
          description: MIME type
          example: text/html
        isProtected:
          type: boolean
          description: Whether note is protected
        blobId:
          type: string
          description: ID of content blob
        dateCreated:
          type: string
          format: date-time
        dateModified:
          type: string
          format: date-time
        utcDateCreated:
          type: string
          format: date-time
        utcDateModified:
          type: string
          format: date-time
    NoteType:
      type: string
      enum:
        - text
        - code
        - render
        - file
        - image
        - search
        - relationMap
        - book
        - noteMap
        - mermaid
        - canvas
        - webView
        - launcher
        - doc
        - contentWidget
        - mindMap
        - geoMap
...
    Revision:
      type: object
      properties:
        revisionId:
          type: string
        noteId:
          type: string
        title:
          type: string
        type:
          $ref: '#/components/schemas/NoteType'
        mime:
          type: string
        dateCreated:
          type: string
          format: date-time
        utcDateCreated:
          type: string
          format: date-time
...